backend/auth-admin/requirements.txt
bcrypt==4.1.2
psycopg2-binary==2.9.9
redis==5.0.11
